Make all methods and states of RepositoryState testable properties
E.g. if the current repository state is SAFE, then the isSafe
property of GitResource and GitRepository will be true. The methods,
like canResetHead is also testable. This applies both for
RepositoriesViewProprtyTester (applicable in the Git Repositories View),
as well as for ResourcePropertiesTester (applicable for resources).

For now, we only declare the properties that we actually use in the
properties testers section in plugin.xml.

Change-Id: I5c72f66ceac7678a03c85794251304a3db82d196