Fix rebase pulldown in toolbar not being accessible during rebase
During rebase, the rebase pulldown menu in the toolbar was always
disabled, resulting in no way to do a "abort" or other actions through
the toolbar.

The problem is that the enabled state of the action handler for the main
action also determines whether the pulldown menu can be opened (the
little arrow besides the icon). So instead of disabling it when rebase
is not possible, also enable it during rebase, and check state again
when executing.

In addition to the above, this fixes RebaseActionHandler's enabled state
to behave the same way as RebaseCurrentRefCommand, in that it's also
enabled with a detached HEAD.

Bug: 387184
JGit-Dependency: I9ec8568fa1100d2e9c8d4ca0e347bf77ec6d8734
Change-Id: Ic69f5c952a49f023c0949f04b3e976be1b267fbe
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>