Configure Branch: Show correct refs when remote is selected
The "merge" key in a branch section of .git/config describes the name of
the branch as it is called on the remote. So it contains e.g.
"refs/heads/master", not "refs/remotes/origin/master".

The dialog for configuring a branch (repo view > Configure Branch...)
just showed all refs, making it likely for users to configure the wrong
thing.

It now calculates the refs based on the value of the remote entry, by
mapping remote tracking refs back to how they are called on the remote
(according to the fetch specs) and only showing the matching refs.

Because the values in the branch combo now depend on the values in the
remote combo, show the remote combo first (for a clear flow).

Bug: 407465
Change-Id: I9c45fa470b6335ddb6a7b1beaf17cc2257f120a1