[historyView] Follow renames by default
With incremental loading of history and showing the renames in file
diffs, I don't know why we shouldn't turn this on by default. I haven't
had a problem with it.

If it really is bad for performance for some users, they can still turn
it off. But the default should be nice.

Change-Id: Id02ca4329d920dc49bf2fedff4ac474c462cb6c1
Signed-off-by: Robin Stocker <robin@nibor.org>