Allow to configure custom test directory
On the EGit Hudson server the home directory is located on a NFS share.
This leads to problems when tests which create test files under the home
directory fail to delete files they have created there due to
limitations of NFS.

Hence allow to configure a custom directory for these tests. Set the
system property "egit.test.tmpdir" to specify a custom directory.

Bug: 414509
Change-Id: If549e4c75d53a63f27d33401385b168bd0f236ff
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>