Fix regression of switched sides in "Compare with Version in Ancestor"
In EGit 3.0, the "new" state was shown on the left and the "old" on the
right, which is consistent with other Eclipse usage, see bug 213780.

In the fix for bug 416808, the sides were switched (probably not
intentionally).

This restores the order and also renames the parameters in the compare
methods to be more clear. It's now also more consistent with how
includeLocal is handled, as local is shown on the left and uses the
"old" parameter for the right side.

Bug: 418029
Change-Id: I4288fac38a3681638601a8fd81f30f5210bdcaa3
Signed-off-by: Robin Stocker <robin@nibor.org>