Update file count on clearing filter
If a filter string in the commit dialog is cleared using the clear
filter button inside the text input control, then the number of files
was not updated.

There seems to be no appropriate listener for filter changes, therefore
I had to add the label update to the refresh job of the tree viewer.

Change-Id: Ic93a9a3043a9efa5be555af5400f951fb1581018
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>