Remove Compare With > History...
It's confusing because it just opens the history view (which can also be
done using Show In > History and Team > Show in History) and then leaves
it to the user to find out how to do a compare. Compare With > Commit...
offers a better alternative by listing the commits in a dialog.

Bug: 421318
Change-Id: I987ab04ea81c3999a339e1f38a3d8f48afbf60a2
Signed-off-by: Robin Stocker <robin@nibor.org>