[sync] Only fetch when synchronizing with remote-tracking branch
For detecting whether a fetch is possible, just check if any of the data
has a dstRemoteName.

Because dstRemoteName was also set for HEAD and local branches (which
does not make much sense), GitSynchronizeData was adapted to only set
them in case of remote-tracking branches. The logic was also improved to
check the remote configs instead of just pattern matching.

I noticed that GitSynchronizeData's getSrcMerge was only used in tests
and getSrcRemoteName only used in PushAction. But there, isEnabled was
testing dstRemoteName, while runPushOperation was using srcRemoteName.
But dstRemoteName makes more sense, as it is used together with
getDstMerge. So, this change also removes the two unused fields,
including associated logic.

Bug: 412973
Change-Id: I87f661fb4292d8e8c40bb8d31640192e25e00c1c
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>