Use Eclipse resource model when checking if a path is ignored
Before, it would iterate over the file system directly. The callers of
the method work with the resource model, which can be out of sync.

This mismatch could cause isIgnored to return false (and the path being
added to .gitignore) because the file system doesn't contain the
directory/file that was asked for.

Bug: 421610
Change-Id: I398e910fcbc39bd36b07d0950184299963f75e62
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>