[historyView] Added "Focus On This File" menu for selected file
Selected file in commit files list have now new "Focus On This File"
context menu, which focuses history view on the selection.

This is a handy option to change the history view focus and also fixes
issue with older Eclipse versions which do not have Show In -> History
View entry.

Bug: 440994
Change-Id: Id31c410e4df24e2959c094fb188ed69a4b04e818
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>