Enable linking with merge tool input
When the merge tool is open, the package explorer "Link with Editor"
action now works when a resource is the input.

A nicer solution would be to have ResourceEditableRevision implement
getAdapter for IFile/IResource and then delegate to that. But that has
strange side effect such as "Save" in the merge tool no longer writing
to the file.

Change-Id: If9f46c4a3e08578de43d2e6e9fc5a2f0e5e4ff13
Signed-off-by: Robin Stocker <robin@nibor.org>