Fix deadlock caused by GitProjectData.logAndUnmapGoneMappedResource
Deadlock could occur as follows:

Thread 1: In GitProjectData.get/load, the class object is locked. In
case a mapped resource is gone, RepositoryProvider.unmap is called,
which waits for job run.

Thread 2: In a job run, GitProvider.getData is called, which waits to
lock the GitProjectData class.

There are also other situations, all of which are caused by
logAndUnmapGoneMappedResource calling RepositoryProvider.unmap. By doing
that in a job where the GitProjectData class is no longer locked, this
should be solved.

Bug: 425973
Change-Id: I74ecc26d696057896ac2790974e9c2e9687c783e
Signed-off-by: Robin Stocker <robin@nibor.org>