Fix ConfigureGerrit Task to respect http.sslVerify config parameter
The ConfigureGerritAfterCloneTask opens a HttpURLConnection to talk to
the REST API of a potential gerrit server. If this is a connection for
the https protocol and if the repository has http.sslVerfiy configured
to false then this task should respect this and accept connections from
servers which certificates can't be validated.

Bug: 456904
Change-Id: I730446d100422e677caa5cd958104666c0833960
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>