Fix that double click is needed to "Check/Uncheck All Repositories"
When using the add existing repository wizard it was necessary to double
click "Check All Repositories"/"Uncheck All Repositories" to
check/uncheck listed repositories. c62ffbf5 tried to fix this but made
it worse for the user: while it unchecked all elements in the UI it then
added all repositories on Finish.

Bug: 450946
Change-Id: I268ff5c3ed2aae2aab0f3bc404410e2fb1ced4ae
Signed-off-by: Dani Megert <Daniel_Megert@ch.ibm.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>