Configurable date format used consistently throughout EGit
Part 2: use the new preference in the reflog and in the interactive
rebase views, and in the remaining label providers.

Follow-up to change Icede623c, which had introduced the preference,
its UI, and had used it in the history view.

With this commit, the date formatting preferences are respected
throughout EGit.

Bug: 477248
Change-Id: Ifda5bc7276c4ba9d22f658c00966d88931beabcf
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>