[historyView] Pre-select single local branch
The actions checkout, delete branch, rename branch and merge most often
are done on a local branch and not on a remote tracking branch. Hence if
the user selects a commit in the history view which has one local branch
only, pre-select this local branch.

Don't list the currently checked out branch for checkout, merge and
delete branch actions as checking out or merging the already checked out
branch has no effect and deleting the current branch isn't allowed.

Change-Id: If494ebfb793d359a5c54da99f7f4a1c900e98873
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>