Use all available space in RebaseInteractiveView
This resolves a TODO in RebaseInteractiveView: if the dynamic columns
need less space than available horizontally, there will be be a large
unused empty space at the right. Can happen if the view is wide, or
commit messages are short, or for instance also in the view is opened
without input (just open it via the quick access).

The solution is to redistribute all available space after pack() using
a TreeColumnLayout with ColumnWeightData and the sizes obtained after
packing.

As a result, the table uses at least all the available horizontal space
and may (as before) use more if the columns don't fit (which then gives
a horizontal scroll bar, also as before).

Change-Id: Ie87961f83339cc34fc0444bcb070cf0aeeba8313
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>