Minor improvement in test for repository deletion
Besides verifying that the shared project has gone from Eclipse when
the repository is deleted also verify that the project is gone on
disk, and (since that project was the only one there) even the whole
working directory is gone.

Instead of just checking for the working tree, do check for each one
explicitly so that in case of a failure it's clearer what exactly
failed.

Change-Id: I8676c0c173ff923ccd799ff9d490c7f8152bbd17
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>