Unify "Open Editor" implementation and behavior for external files
Repository view used a different but very similar implementation for
opening editors as Staging view. The difference is that the Staging view
always opened default text editor on external files, and Repository view
allows IDE decide which editor is more suitable. The later one is better
because user will usually get the right editor for the file (e.g. Java
editor for *.java with syntax coloring etc) and also can configure which
editor should be used for specific file type.

Change-Id: I756fbb5f491b7ae47b01789bdca6faec08d6269b
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>