Catch IOException in ConfigureGerritAfterCloneTask
This task must not make the git clone be reported as failed -- the
repository was already successfully cloned. Catch the exception and
return false, which is the safe thing to do: in the worst case, the
user may have to do the Gerrit config setup him- or herself via the
ConfigureGerritWizard.

Also clean up closing streams and disconnecting.

Bug: 494342
Change-Id: Ie10cb12da8e8d732c956e7352c15702898e23710
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>