Speed up the IgnoreOperation
Adding files to .gitignore would take a lot of time if many files were
selected. Improve this by:

* grouping paths by folder,
* not checking each file individually for eligibility via
  RepositoryUtil.canBeAutoIgnored() but doing a batch check in a
  single tree walk for all files in a folder,
* not adding them one by one to the .gitignore file(s) but batch adding
  per folder.
* running the modifications in a WorkspaceJob to batch resource update
  events for .gitignore files in the workspace.

This makes it possible to even use the "Ignore" action in the staging
view on a (non-expanded) folder containing 50000 unstaged files. This
is now snappy.

The following IndexDiff recomputation is very slow, though. On my Mac
it takes about 40 seconds until the staging view and the decorations in
the package explorer update. This is not fixed by this change, and it's
unclear to me how this could be fixed. A .gitignore file with 50000
single-file entries slows down JGit tremendously because it will
process all those match rules for all files in that directory, which
gives quadratic behavior.

This change therefore also adds two measures to mitigate this:

* a warning dialog explaining this, giving the user a chance to back
  out. The threshold for this warning dialog to appear has been set
  somewhat arbitrarily at 500 files to be ignored.
* a new menu action "Ignore Folders" to ignore folders instead of the
  files contained in those folders. This action is only available if
  only folders are selected in the staging view.
  
With this new action, it's possible to ignore that huge folder also from
the staging view. Before, it could be ignored only via the Team->Ignore
context menu action in the package explorer.

Bug: 500106
Change-Id: Icf2c9021dd20ec1372489cfe2f4f0da446a6a78c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>