Fix recursive adaptation
The typo fix done in commit ddb2f97e uncovered an endless recursion
problem. Remove that adapter definition completely for now. It was
introduced as part of a change that made the history view react on
working set selections, but for that, the adapter definition for
IWorkingSet is sufficient.

I'll have to re-visit this later to see more precisely if there's a
simple way to have this adapter definition without running into an
endless recursion.

Bug: 506151
Change-Id: I83720a90665701d130e2c70039ac73624b12bba2
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>