Do not auto-fill reference in pull wizard with local branch name
The reference field was auto-filled with the name of the currently
checked out local branch. This doesn't make sense since there is no
guarantee that a branch with the same name exists upstream. Checking if
this is the case would be too slow since it involves a remote request.
Hence leave the reference field empty, the user can use content assist
to find candidate upstream references.

Change-Id: I624cd7bd095830455910e9f212a13fb48ebea6fc
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>