Don't steal the focus in the find widget of the history view
The find toolbar widget of the history view does some ugly focus
juggling. Make sure to do this only initially, or when the widget is
explicitly shown by the user triggering the findToolbarAction, but not
when the control is re-created in response to other events, such as
switching between history pages.

This focus back-and-forth is a hack, but unfortunately it's the only
way I have found to fix the wrong background of the whole toolbar when
showing the find widget makes the toolbar move from right of the tabs
to below the tabs.

Bug: 507484
Change-Id: I96c2cadeb988612371dceeddc2f8967cc9566e1a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>